snapmgr/requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.4.4
MarkupSafe==0.23
pyodbc==3.0.10
python-ldap==2.4.20
Werkzeug==0.10.4
